Gunicorn reload on knowledgebase changes

+++ kernel/app.py
@@ -33,10 +33,14 @@ def query():
 
 @app.route("/deploy", methods=['PUT'])
 def deploy():
-    shutil.rmtree(os.path.join(APP_KB,"concepts"))
-    shutil.rmtree(os.path.join(APP_KB,"hierarchy"))
-    shutil.rmtree(os.path.join(APP_KB,"relations"))
-    shutil.rmtree(os.path.join(APP_KB,"rules"))
+    if Path(os.path.join(APP_KB,"concepts")).exists:
+        shutil.rmtree(os.path.join(APP_KB,"concepts"))
+    if Path(os.path.join(APP_KB,"hierarchy")).exists:
+        shutil.rmtree(os.path.join(APP_KB,"hierarchy"))
+    if Path(os.path.join(APP_KB,"relations")).exists:
+        shutil.rmtree(os.path.join(APP_KB,"relations"))
+    if Path(os.path.join(APP_KB,"rules")):
+        shutil.rmtree(os.path.join(APP_KB,"rules"))
     for filePath in request.files.keys():
         directoryPath = os.path.dirname(os.path.join(APP_KB, filePath))
         Path(directoryPath).mkdir(parents=True, exist_ok=True)
